+++ src/main/java/com/dxfeed/event/market/EventMapper.java
@@ -1,3 +1,6 @@
+// Copyright (c) 2024 Devexperts LLC.
+// SPDX-License-Identifier: MPL-2.0
+
 package com.dxfeed.event.market;
 
 import com.dxfeed.event.EventType;
@@ -8,39 +11,42 @@ import org.graalvm.word.WordFactory;
 public abstract class EventMapper<V extends EventType<?>, T extends DxfgEventType>
     extends Mapper<V, T> {
 
-  public T toNative(final V jObject) {
-    if (jObject == null) {
+  public T toNative(final V javaObject) {
+    if (javaObject == null) {
       return WordFactory.nullPointer();
     }
-    final T nObject = createNativeObject();
-    fillNative(jObject, nObject, true);
-    return nObject;
+
+    final T nativeObject = createNativeObject();
+
+    fillNative(javaObject, nativeObject, true);
+
+    return nativeObject;
   }
 
   protected abstract T createNativeObject();
 
-  public T toNativeObjectWithCast(final EventType<?> jEvent) {
-    return toNative((V) jEvent);
+  public T toNativeObjectWithCast(final EventType<?> javaEvent) {
+    return toNative((V) javaEvent);
   }
 
-  public V toJavaObjectWithCast(final DxfgEventType nEvent) {
-    return toJava((T) nEvent);
+  public V toJavaObjectWithCast(final DxfgEventType nativeEvent) {
+    return toJava((T) nativeEvent);
   }
 
-  public void fillNativeObjectWithCast(final EventType<?> jEvent, final DxfgEventType nEvent) {
-    fillNative((V) jEvent, (T) nEvent, true);
+  public void fillNativeObjectWithCast(final EventType<?> javaEvent, final DxfgEventType nativeEvent) {
+    fillNative((V) javaEvent, (T) nativeEvent, true);
   }
 
-  public void cleanNativeObjectWithCast(final DxfgEventType nEvent) {
-    cleanNative((T) nEvent);
+  public void cleanNativeObjectWithCast(final DxfgEventType nativeEvent) {
+    cleanNative((T) nativeEvent);
   }
 
-  public void fillJavaObjectWithCast(final DxfgEventType nEvent, final EventType<?> jEvent) {
-    fillJava((T) nEvent, (V) jEvent);
+  public void fillJavaObjectWithCast(final DxfgEventType nativeEvent, final EventType<?> javaEvent) {
+    fillJava((T) nativeEvent, (V) javaEvent);
   }
 
-  public void releaseWithCast(final DxfgEventType nEvent) {
-    release((T) nEvent);
+  public void releaseWithCast(final DxfgEventType nativeEvent) {
+    release((T) nativeEvent);
   }
 
   public abstract T createNativeObject(final String symbol);